+++ src/core/SlashCommandHandler.js
@@ -3,20 +3,20 @@ import Slack from '../integrations/Slack';
 const slack = new Slack();
 
 export default class SlashCommandHandler {
-  async teams(req, res, next) {
+  async feedback(req, res, next) {
     try {
       var actions = [];
       if (req.user && req.user.is_sims_facilitator) {
         actions.push({
           name: 'landing_page_menu',
-          text: 'Create Team...',
+          text: 'Record Feedback...',
           style: 'primary',
           type: 'button',
-          value: 'create_team'
+          value: 'record_full_feedback'
         });
         actions.push({
           name: 'landing_page_menu',
-          text: 'Analytics',
+          text: 'Analytics...',
           type: 'button',
           value: 'analytics'
         });
@@ -25,7 +25,7 @@ export default class SlashCommandHandler {
         name: 'landing_page_menu',
         text: 'Help',
         type: 'button',
-        url: 'https://github.com/andela-stuff/andela-teams-slack/wiki'
+        url: 'https://github.com/andela/feedback-bot/wiki'
       });
       const attachments = [{
         callback_id: 'landing_page_menu',