+++ src/server.js
@@ -6,9 +6,7 @@ import bodyParser from 'body-parser';
 
 import FeedbackAnalytics from './analytics/FeedbackAnalytics';
 import DataHandler from './core/DataHandler';
-import Eventhandler from './core/EventHandler';
 import InteractionHandler from './core/InteractionHandler';
-import PivotalTrackerAnalytics from './analytics/PivotalTrackerAnalytics';
 import SlashCommandHandler from './core/SlashCommandHandler';
 import Utility from './core/Utility';
 
@@ -22,33 +20,21 @@ app.use(bodyParser.urlencoded({extended: false}));
 app.use(bodyParser.json());
 
 const data = new DataHandler();
-const event = new Eventhandler();
 const feedbackAnalytics = new FeedbackAnalytics();
 const interaction = new InteractionHandler();
-const ptAnalytics = new PivotalTrackerAnalytics();
 const slash = new SlashCommandHandler();
 const utils = new Utility();
 
 app.get('/', async (req, res) => {
-  res.status(200).send("Hello World!<br /><br />Welcome to Andela Teams for Slack.");
+  res.status(200).send("Hello World!<br /><br />Welcome to Feedback Bot for Slack.");
 });
 
 app.get('/api/analytics/feedback/:token',
   feedbackAnalytics.get);
 
-app.get('/api/analytics/pt/:token',
-  ptAnalytics.get);
-
 app.post('/data/external',
   utils.getUserObjectFromReqBodyPayloadUserId,
   data.dialogSuggestions);
-
-app.post('/events', 
-  event.challenge,
-  utils.getUserObjectFromReqBodyEventUser,
-  utils.rejectUsersWithNoEmailOrGithub,
-  event.addMeReaction,
-  utils.handleErrors);
 
 app.post('/interactions',
   utils.postEmptyMessage,
@@ -60,11 +46,11 @@ app.post('/interactions',
   interaction.messageAction,
   utils.handleErrors);
 
-app.post('/slash/teams',
+app.post('/slash/feedback-bot',
   utils.postWelcomeMessage,
   utils.getUserObjectFromReqBodyUserId,
   utils.rejectUsersWithNoEmailOrGithub,
-  slash.teams,
+  slash.feedback,
   utils.handleErrors);
 
 let server = app.listen(process.env.PORT || 5000, () => {